LAU-020 Add login function to login input line
